requirements.txt
tensorflow>=2.8.0
numpy>=1.21.0
matplotlib>=3.5.0
pillow>=9.0.0
scipy>=1.7.0
scikit-image>=0.19.0
tqdm>=4.62.0 